销售出库单支持负库存的代码。。
